requirements.txt
Flask==3.0.3
numpy==2.0.0
pydantic==2.7.4
pytest==8.2.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
psycopg2-binary==2.9.9
python-dotenv==1.0.1
Flask-JWT-Extended==4.6.0
celery==5.4.0
redis==5.0.8
openai==1.40.3
